05_redis/requirements.txt
fastapi
pydantic
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pyjwt
python-multipart
pymongo
redis


# pandas
# prometheus_client
# requests
# pika
# confluent-kafka
